Train model

combines both training steps

Usage

trainSWR(
  ts_input,
  ts_output,
  iter = 10,
  runs = 10,
  log = FALSE,
  parallel = TRUE,
  return = "best",
  param_selection = "best_rmse"
)

Arguments

ts_input

a vector or ts object containing the input time series

ts_output

a vector or ts object (on the same time scale as ts_input) containing the target time series

iter

number of iterations (maximum number of windows)

runs

number of independent model runs

log

whether a log-linear model should be used

parallel

should the runs be computed in parallel? If FALSE, all runs are computed in serial. If TRUE, all runs are computed in parallel with a maximum number of cores. If a scalar is provided, the number of cores is set manually.

return

either "best" (best model run is returned), or "all" (all model runs are returned)

param_selection

either "max" (maximum number of windows), or "best_rmse", "best_aic", or "best_bic" to optimize RMSE, AIC, or BIC, respectively

Examples

# train a model based on one year of observations
set.seed(42)
data(sampleWatershed)
trainSWR(sampleWatershed$rain[1:365],
         sampleWatershed$gauge[1:365],
         iter = 2,
         runs = 1,
         parallel = FALSE)
#> $param
#>          delta     sigma
#> [1,] 0.9206601 0.4795775
#> 
#> $mix
#> [1] 0.8067206
#> 
#> $train_hist
#>     operation      aic      bic     rmse    nrmse        r2       nse       kge
#> 1 iteration 1 2064.442 2076.141 4.058624 1.113781 0.7141318 0.7141318 0.7847140
#> 2 iteration 2 2070.600 2094.000 4.059506 1.114023 0.7140076 0.7140076 0.7845635
#> 
#> $train_metrics
#>        aic      bic     rmse    nrmse        r2       nse      kge
#> 1 2064.442 2076.141 4.058624 1.113781 0.7141318 0.7141318 0.784714
#> 
#> $param_hist
#> $param_hist[[1]]
#> $param_hist[[1]]$mix
#> [1] 0.8067206
#> 
#> $param_hist[[1]]$param
#>          delta     sigma
#> [1,] 0.9206601 0.4795775
#> 
#> 
#> $param_hist[[2]]
#> $param_hist[[2]]$mix
#> [1] 0.3821931 0.4233929
#> 
#> $param_hist[[2]]$param
#>          delta     sigma
#> [1,] 0.5371369 0.1421062
#> [2,] 1.3132182 0.2315104
#> 
#> 
#> 
#> $fitted
#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 6.217822e-02 2.801525e-02
#>   [6] 3.997532e-03 3.322960e-03 8.050557e-03 2.693861e-02 3.155861e-02
#>  [11] 1.004525e-02 2.417484e-02 9.643578e-02 3.581398e-01 1.229584e+00
#>  [16] 1.163768e+00 3.110009e+00 5.179173e+00 7.613242e+00 4.920041e+00
#>  [21] 1.270459e+00 4.139583e+00 1.173153e+01 7.942569e+00 2.266031e+01
#>  [26] 2.096067e+01 1.416891e+01 9.908174e+00 2.898634e+00 8.902383e-01
#>  [31] 1.017245e+00 2.808820e-01 7.277854e-01 1.506158e+00 1.561847e+00
#>  [36] 1.853365e+00 1.205955e+00 2.149353e-01 1.022640e-02 2.489204e-03
#>  [41] 9.275757e-03 1.557913e-02 5.337593e-02 2.615599e-02 2.195302e-02
#>  [46] 8.557063e-01 4.073441e+00 5.663944e+00 4.268616e+00 1.524832e+00
#>  [51] 7.059932e-01 1.354724e+00 6.159217e+00 7.221264e+00 3.316016e+00
#>  [56] 1.498110e+00 7.999304e-01 3.009258e-01 4.373660e-02 8.364388e-01
#>  [61] 3.543638e+00 6.261602e+00 1.529745e+01 1.316316e+01 2.556855e+01
#>  [66] 1.009289e+01 2.929657e+00 1.043934e+00 1.010695e+00 2.332934e+00
#>  [71] 4.510038e+00 3.150664e+00 1.010510e+01 2.809174e+01 5.450037e+01
#>  [76] 3.281317e+01 1.302312e+01 4.144281e+01 4.130922e+01 2.088576e+01
#>  [81] 1.138322e+01 1.116797e+01 4.889296e+00 2.417766e+00 7.114288e+00
#>  [86] 6.479581e+00 6.056344e+00 4.316496e+00 1.345737e+00 1.837100e+00
#>  [91] 2.408744e+00 8.385303e+00 1.370062e+01 8.300230e+00 6.575473e+00
#>  [96] 2.465078e+00 9.605378e-01 1.196773e-01 5.200720e-04 0.000000e+00
#> [101] 0.000000e+00 0.000000e+00 0.000000e+00 4.044620e+00 1.492594e+01
#> [106] 5.245142e+00 1.288197e+01 1.114103e+01 2.902884e+00 2.376116e-01
#> [111] 1.012915e-03 0.000000e+00 7.935079e-02 2.957556e-01 7.927749e-02
#> [116] 1.330845e-01 2.111879e-01 3.229832e-02 1.412839e-04 0.000000e+00
#> [121] 0.000000e+00 0.000000e+00 0.000000e+00 4.505349e+00 2.162744e+01
#> [126] 2.318298e+01 1.082931e+01 4.637638e+00 9.000410e+00 4.326391e+00
#> [131] 5.997247e-01 5.344079e-02 5.469655e-02 2.880572e-02 1.563386e-02
#> [136] 1.506634e-02 4.795278e-02 7.782539e-03 6.843867e-01 4.148992e+00
#> [141] 8.463648e+00 9.840513e+00 6.176402e+00 1.919582e+00 3.302900e-01
#> [146] 2.021127e-01 1.503035e+00 7.489635e+00 2.461895e+01 3.906995e+01
#> [151] 1.946235e+01 5.896179e+00 1.859448e+00 1.785614e+00 2.749668e+00
#> [156] 1.149660e+00 8.515279e-01 1.776677e-01 4.469972e-02 1.341375e-01
#> [161] 1.019783e+00 4.169678e+00 5.080062e+00 5.526248e+00 1.386474e+01
#> [166] 1.361265e+01 8.414872e+00 5.268275e+00 9.481112e+00 8.986648e+00
#> [171] 2.613449e+00 2.964771e+00 1.441533e+00 8.541055e-01 1.916170e+00
#> [176] 5.134246e-01 4.569783e-01 1.628393e+00 2.238380e+00 2.600704e+00
#> [181] 4.383173e+00 2.360763e+00 8.497750e-01 5.301780e-01 9.173646e-02
#> [186] 5.975993e-02 5.089414e-01 2.643517e+00 6.687637e+00 4.801632e+00
#> [191] 3.137497e+00 6.929800e+00 6.995353e+00 1.585397e+00 1.483064e-01
#> [196] 5.700407e-01 3.252765e+00 6.376802e+00 2.357213e+00 1.142778e+00
#> [201] 4.638068e+00 7.867767e+00 7.997432e+00 1.412953e+00 5.183836e-02
#> [206] 1.507253e-01 6.548392e-01 6.351006e-01 6.122920e-01 4.276944e-01
#> [211] 2.636144e-01 7.886271e-01 3.182346e-01 4.191015e-02 3.372181e-02
#> [216] 2.776977e-02 1.335319e-01 7.227211e-01 1.065555e+00 5.910775e-01
#> [221] 8.818014e-01 1.541530e+00 6.185366e-01 1.507164e-01 1.454549e-01
#> [226] 5.018146e-01 1.239935e+00 3.113743e+00 1.257624e+00 1.628594e-01
#> [231] 1.981922e-02 1.709957e+00 7.214399e+00 4.745730e+00 1.721317e+00
#> [236] 1.199761e+00 7.411549e-01 1.194424e+00 1.596933e+00 4.879275e+00
#> [241] 3.742141e+00 8.213965e-01 2.068679e-01 2.448405e-01 7.365383e-01
#> [246] 1.296448e+00 1.580919e+00 1.961478e+00 2.824332e+00 1.148759e+00
#> [251] 3.500575e+00 1.126185e+01 7.975282e+00 3.343674e+00 2.154669e+00
#> [256] 6.259189e-01 1.214702e-01 2.781760e-02 4.373532e-02 1.430868e+00
#> [261] 5.000656e+00 1.074274e+00 3.710270e-01 6.563127e-02 2.077067e-01
#> [266] 8.649504e-01 1.247017e+00 2.998089e+00 8.803901e+00 8.729925e+00
#> [271] 3.743377e+00 7.814347e-01 1.436428e-01 1.733288e-01 1.983671e-01
#> [276] 1.166820e+00 6.404646e+00 1.103271e+01 4.012391e+00 1.143173e+00
#> [281] 1.820101e-01 1.635802e-02 1.038702e-01 8.848010e-01 2.346634e+00
#> [286] 1.558965e+00 5.583712e-01 1.142212e+00 8.434673e-01 9.941152e-01
#> [291] 5.012389e-01 4.593364e-01 1.362438e+00 1.382386e+00 2.093366e-01
#> [296] 9.988695e-03 3.567153e-02 4.005893e-02 1.745886e-02 4.030745e-03
#> [301] 1.636389e-03 4.844906e-02 1.775729e-01 4.411859e-02 1.338848e-02
#> [306] 1.261665e-01 4.108412e-01 3.753702e-01 1.491290e-01 2.310677e-01
#> [311] 2.314027e-01 1.596474e-01 1.323571e-01 7.561581e-02 1.996525e-01
#> [316] 1.194709e-01 2.101872e-01 6.380543e-02 1.456915e-01 3.886655e-01
#> [321] 5.177355e-01 2.873337e+00 5.082728e+00 1.217411e+00 1.745870e-01
#> [326] 2.573305e-02 2.864296e-02 1.352432e-01 2.305033e-01 2.558987e-01
#> [331] 1.075933e+00 1.914044e+00 1.192089e+00 2.471272e-01 2.729169e-01
#> [336] 1.262967e+00 5.379778e+00 7.808577e+00 1.718393e+00 3.120807e-01
#> [341] 3.079686e-01 1.894380e+00 5.709216e+00 1.543735e+00 3.472009e-01
#> [346] 6.934855e-02 6.947025e-02 4.596227e-01 9.776968e-01 5.952126e-01
#> [351] 1.382051e-01 1.184379e-02 2.903563e-03 4.324500e-01 4.460532e+00
#> [356] 1.128999e+01 3.231890e+00 1.515744e+00 1.766390e+00 3.158764e-01
#> [361] 1.123217e-02 2.791050e-03 2.918083e-01 2.036096e+00 6.144832e+00
#> 
#> $residuals
#>   [1]  3.042305e-01  2.866576e-01  2.661559e-01  1.867709e-01  2.103170e-01
#>   [6]  2.273787e-01  2.225618e-01  2.130749e-01  1.916241e-01  1.826109e-01
#>  [11]  1.993649e-01  1.830387e-01  1.107778e-01 -1.465330e-01 -1.022371e+00
#>  [16] -9.452053e-01 -2.871677e+00 -4.906061e+00 -7.089717e+00 -4.367227e+00
#>  [21] -6.737130e-01 -2.356668e+00 -9.974241e+00 -6.514773e+00 -1.094506e+01
#>  [26] -4.522703e+00 -3.039423e+00 -3.940717e+00  9.454337e-01  1.694440e+00
#>  [31]  1.204992e+00  1.743660e+00  3.961468e-01 -2.907002e-01 -3.756773e-01
#>  [36] -6.379078e-01 -9.300613e-02  9.565902e-01  1.007536e+00  9.310701e-01
#>  [41]  8.437412e-01  8.191327e-01  7.593698e-01  7.536407e-01  7.834707e-01
#>  [46] -5.760463e-02 -3.300966e+00 -4.810927e+00 -3.217904e+00 -5.253742e-01
#>  [51]  2.714984e-01 -3.223172e-01 -5.060912e+00 -5.903298e+00 -1.880898e+00
#>  [56]  2.811443e-01  1.092815e+00  1.511278e+00  1.384060e+00  4.961714e-01
#>  [61] -2.309876e+00 -4.998551e+00 -1.356945e+01 -1.056750e+01  1.049247e+01
#>  [66] -6.170864e-02  3.989665e+00  4.008269e+00  2.760153e+00  4.421166e-01
#>  [71] -1.775258e+00 -6.501895e-01 -7.655882e+00  1.891993e+00  2.311319e+01
#>  [76] -1.694528e+00 -1.014982e+00  2.518770e+01 -3.234640e+00  1.650899e-01
#>  [81]  5.640513e+00  1.975078e+00  1.920196e+00  3.769352e+00 -1.220051e+00
#>  [86] -3.290733e-01  2.437215e+00  2.602825e+00  4.621720e+00  2.922222e+00
#>  [91]  2.313967e+00 -3.662592e+00 -7.696554e+00 -1.088026e+00  1.002832e+00
#>  [96]  2.843396e+00  3.542513e+00  4.017272e+00  4.063209e+00  4.027119e+00
#> [101]  3.269288e+00  2.932475e+00  4.136949e+00  2.912419e+01  5.356094e+00
#> [106]  8.687565e-01 -9.836678e-01 -2.025094e+00  4.565590e+00  6.498659e+00
#> [111]  5.087801e+00  4.393220e+00  3.618276e+00  2.819770e+00  2.941061e+00
#> [116]  2.850644e+00  2.754236e+00  2.925803e+00  2.932333e+00  1.991593e+00
#> [121]  1.965966e+00  1.940339e+00  1.951322e+00 -1.993892e+00 -4.676933e+00
#> [126]  5.372950e+00  7.988317e+00  8.578633e+00  4.032810e+00  7.535304e+00
#> [131]  6.832139e+00  6.792661e+00  5.290388e+00  4.876957e+00  4.011485e+00
#> [136]  3.755781e+00  2.972386e+00  2.906387e+00  2.090664e+00 -1.286077e+00
#> [141] -5.348122e+00  5.201642e-01  1.438513e+00  3.498723e+00  4.026320e+00
#> [146]  3.678565e+00  1.630796e+00 -2.437432e+00 -1.041998e+00 -5.351989e+00
#> [151]  4.920027e+00  1.032213e+01  8.354789e+00  5.646250e+00  1.570333e+00
#> [156]  3.865933e+00  3.102370e+00  3.703010e+00  2.986622e+00  2.216235e+00
#> [161]  1.023064e+00 -2.071916e+00 -2.949350e+00 -3.417502e+00 -1.158393e+01
#> [166] -8.414006e+00 -4.058262e+00 -1.350987e+00 -1.097383e+00  7.516569e-01
#> [171]  5.697060e+00  4.393873e+00  5.221518e+00  5.552674e+00  4.161118e+00
#> [176]  4.538779e+00  3.972852e+00  2.215675e+00  1.257892e+00  1.389804e+00
#> [181] -6.123251e-01  1.080592e+00  2.295039e+00  1.882432e+00  1.801009e+00
#> [186]  1.536444e+00  1.211736e+00 -3.956530e-01 -4.644789e+00 -2.520818e+00
#> [191]  9.628419e-01 -6.694609e-01 -1.906540e+00  3.137314e+00  3.366270e+00
#> [196]  2.413688e+00  8.109644e-01 -1.800531e+00  1.274516e+00  1.815324e+00
#> [201] -1.409052e+00 -3.804038e+00 -4.189975e+00  2.141894e+00  2.968501e+00
#> [206]  2.360732e+00  1.556415e+00  1.506594e+00  1.273132e+00  1.208780e+00
#> [211]  1.241064e+00  5.952373e-01  9.960704e-01  1.195514e+00  1.086549e+00
#> [216]  9.790099e-01  8.256545e-01  2.694145e-01 -1.393176e-01  3.022106e-01
#> [221] -1.780139e-02 -7.141400e-01  1.942092e-01  5.888090e-01  5.611213e-01
#> [226]  1.901176e-01 -5.113923e-01 -2.366896e+00 -5.730143e-01  4.741575e-01
#> [231]  5.805876e-01 -1.127855e+00 -6.518806e+00 -3.925662e+00 -9.817915e-01
#> [236] -5.883711e-01 -1.663752e-01 -5.940174e-01 -1.036797e+00 -4.293513e+00
#> [241] -2.984310e+00 -1.001761e-01  4.777422e-01  3.994985e-01 -1.361316e-01
#> [246] -7.472956e-01 -1.072038e+00 -1.412325e+00 -2.242231e+00 -4.531662e-01
#> [251] -2.742745e+00 -1.027704e+01 -6.628027e+00 -2.120895e+00 -1.056364e+00
#> [256]  3.552337e-01  7.535129e-01  7.776061e-01  7.360613e-01 -7.133092e-01
#> [261] -4.330689e+00 -4.519008e-01  2.183967e-01  4.835213e-01  2.792085e-01
#> [266] -4.036622e-01 -8.040337e-01 -2.511174e+00 -7.991155e+00 -7.898874e+00
#> [271] -3.036801e+00 -1.920110e-01  3.725606e-01  3.062644e-01  2.446160e-01
#> [276] -7.641082e-01 -5.917731e+00 -1.038837e+01 -3.053204e+00 -2.901562e-01
#> [281]  5.977866e-01  6.169979e-01  4.452824e-01 -3.100213e-01 -1.746227e+00
#> [286] -9.365916e-01  5.301861e-02 -5.601106e-01 -2.943147e-01 -4.852338e-01
#> [291] -5.093385e-02 -3.465844e-02 -9.633872e-01 -9.906570e-01  1.714091e-01
#> [296]  3.542825e-01  2.952844e-01  2.755207e-01  2.904327e-01  3.012981e-01
#> [301]  2.883162e-01  2.210018e-01  8.418981e-02  2.048306e-01  2.355607e-01
#> [306]  1.202199e-01 -1.670175e-01 -1.341092e-01  8.444388e-02 -2.620264e-03
#> [311] -1.064335e-02  5.342376e-02  7.558864e-02  1.272045e-01  8.293247e-03
#> [316]  7.566133e-02 -2.530583e-02  1.133878e-01  7.491945e-04 -2.191604e-01
#> [321] -3.507931e-01 -2.711520e+00 -4.895284e+00 -1.040218e+00  4.348846e-05
#> [326]  1.463347e-01  1.459876e-01  4.451276e-02 -5.587279e-02 -8.383090e-02
#> [331] -8.987394e-01 -1.739414e+00 -1.020021e+00 -7.762212e-02 -1.008491e-01
#> [336] -1.085773e+00 -5.156456e+00 -7.595506e+00 -1.518135e+00 -1.143858e-01
#> [341] -1.000229e-01 -1.655682e+00 -5.460267e+00 -1.310162e+00 -1.315671e-01
#> [346]  1.385972e-01  1.333501e-01 -2.593650e-01 -7.697510e-01 -3.975177e-01
#> [351]  6.205249e-02  1.807257e-01  1.871032e-01 -2.450059e-01 -4.285902e+00
#> [356] -1.113074e+01 -3.064947e+00 -1.339649e+00 -1.577481e+00 -1.141543e-01
#> [361]  2.033034e-01  2.245581e-01 -4.798460e-02 -1.787147e+00 -5.854879e+00
#> 
#> attr(,"class")
#> [1] "SWR"